requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
dnspython==2.7.0
idna==3.10
lxml==5.3.0
pymongo==4.10.1
RateMyProfessorAPI==1.3.6
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
